complete shopping pages
